Practice and small updates

+++ week-02/day-1/C++ Basics/Exercise_9.cpp
@@ -4,7 +4,7 @@
 using namespace std;
 
 int main() {
-	int i = 1357988018575474;
+  long long int i = 1357988018575474;
 	int d = 11;
 	if ( i % d == 0 )
 		cout << d << " is a divisor of " << i <<".";
